Added x5t generation when importing a cert and extra fields
The extra fields allow non-JSON-string imported keys (for instance PEM) to have additional JWK fields
added on to the fields captured from the import.  In particular, fields like a corresponding x5t can
be added to a private RSA key so that tokens generated in the presence of the public x509 cert can
use that as a key identifier instead of or in addition to kid.